# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Gameng API
  version: 0.0.1
  description: Server-side data-driven RPG engine API
servers:
  - url: http://localhost:3000
...
paths:
...
  /{gameInstanceId}/tx:
    post:
      operationId: submitTransaction
      summary: Submit a transaction
      description: >
        Submit an atomic command to mutate game state. Returns 200 with the
        result regardless of whether the transaction was accepted or rejected.
        Domain errors are reported via `accepted: false` and `errorCode` in the
        response body. Returns 400 only for malformed or unparseable request bodies.
        CreateActor requires the ADMIN_API_KEY via Bearer token; all other
        transaction types require a valid actor Bearer token.
      tags:
        - transaction
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/gameInstanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Transaction"
            example:
              txId: tx_001
              type: CreatePlayer
              gameInstanceId: instance_001
              playerId: player_1
      responses:
        "200":
          description: Transaction processed (check `accepted` field for outcome)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResult"
              examples:
                accepted:
                  summary: Transaction accepted
                  value:
                    txId: tx_001
                    accepted: true
                    stateVersion: 1
                rejected:
                  summary: Transaction rejected
                  value:
                    txId: tx_002
                    accepted: false
                    stateVersion: 0
                    errorCode: PLAYER_NOT_FOUND
                    errorMessage: "Player 'player_99' not found."
                unsupported_type:
                  summary: Unsupported transaction type
                  value:
                    txId: tx_003
                    accepted: false
                    stateVersion: 0
                    errorCode: UNSUPPORTED_TX_TYPE
                    errorMessage: "Transaction type 'DeletePlayer' is not yet supported."
        "400":
          description: Malformed request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid Bearer token. For CreateActor, the ADMIN_API_KEY is required. For all other types, a valid actor token is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                missing_token:
                  summary: No Authorization header
                  value:
                    errorCode: UNAUTHORIZED
                    errorMessage: "Missing or invalid Bearer token."
                invalid_admin_key:
                  summary: CreateActor with non-admin token
                  value:
                    errorCode: UNAUTHORIZED
                    errorMessage: "Missing or invalid admin API key."
        "404":
          description: Game instance not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                errorCode: INSTANCE_NOT_FOUND
                errorMessage: "Game instance 'unknown' not found."
...
components:
  parameters:
    gameInstanceId:
      name: gameInstanceId
      in: path
      required: true
      description: Unique identifier for the game instance
      schema:
        type: string
        minLength: 1
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - errorCode
        - errorMessage
      properties:
        errorCode:
          type: string
          minLength: 1
        errorMessage:
          type: string

    # -- Config domain --
...
    Transaction:
      type: object
      required:
        - txId
        - type
        - gameInstanceId
      additionalProperties: false
      properties:
        txId:
          type: string
          minLength: 1
          description: Unique transaction identifier (for idempotency).
        type:
          type: string
          enum:
            - CreateActor
            - CreatePlayer
            - CreateCharacter
            - CreateGear
            - LevelUpCharacter
            - LevelUpGear
            - EquipGear
            - UnequipGear
        gameInstanceId:
          type: string
          minLength: 1
        playerId:
          type: string
          minLength: 1
        actorId:
          type: string
          minLength: 1
          description: Required for CreateActor.
        apiKey:
          type: string
          minLength: 1
          description: Required for CreateActor.
        classId:
          type: string
          minLength: 1
          description: Required for CreateCharacter.
        characterId:
          type: string
          minLength: 1
          description: Required for CreateCharacter, LevelUpCharacter, EquipGear. Optional hint for UnequipGear.
        gearDefId:
          type: string
          minLength: 1
          description: Required for CreateGear.
        gearId:
          type: string
          minLength: 1
          description: Required for LevelUpGear, EquipGear, UnequipGear.
        slotPattern:
          type: array
          items:
            type: string
            minLength: 1
          minItems: 1
          description: Optional for EquipGear — which occupation pattern to use.
        swap:
          type: boolean
          description: Optional for EquipGear — if true, use swap mode instead of strict mode.
        levels:
          type: integer
          minimum: 1
          description: Optional for LevelUpCharacter / LevelUpGear — number of levels to gain. Default 1.
      allOf:
        - if:
            not:
              properties:
                type:
                  const: CreateActor
              required:
                - type
          then:
            required:
              - playerId
        - if:
            properties:
              type:
                const: CreateActor
            required:
              - type
          then:
            required:
              - actorId
              - apiKey
        - if:
            properties:
              type:
                const: CreateCharacter
            required:
              - type
          then:
            required:
              - classId
              - characterId
        - if:
            properties:
              type:
                const: CreateGear
            required:
              - type
          then:
            required:
              - gearDefId
              - gearId
        - if:
            properties:
              type:
                const: LevelUpCharacter
            required:
              - type
          then:
            required:
              - characterId
        - if:
            properties:
              type:
                const: LevelUpGear
            required:
              - type
          then:
            required:
              - gearId
        - if:
            properties:
              type:
                const: EquipGear
            required:
              - type
          then:
            required:
              - gearId
              - characterId
        - if:
            properties:
              type:
                const: UnequipGear
            required:
              - type
          then:
            required:
              - gearId
    TransactionResult:
      type: object
      required:
        - txId
        - accepted
        - stateVersion
      additionalProperties: false
      properties:
        txId:
          type: string
          minLength: 1
        accepted:
          type: boolean
        stateVersion:
          type: integer
          minimum: 0
          description: State version after processing (same as before if rejected).
        errorCode:
          type: string
          minLength: 1
          description: Present when accepted is false.
        errorMessage:
          type: string
          description: Human-readable error description.
      if:
        properties:
          accepted:
            const: false
        required:
          - accepted
      then:
        required:
          - errorCode

    # -- New response schemas --
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        Bearer token for authentication. For CreateActor transactions, use the
        ADMIN_API_KEY. For all other endpoints, use the actor's apiKey obtained
        via CreateActor.